+++ lib/src/track.dart
@@ -7,55 +7,16 @@ import 'package:speedvector7drl/src/road.dart';
 
 class Track {
   final int _width = 32;
-  final _height = 32;
+  final int _height = 32;
   List<String> _road;
-  int _roadWidth;
-  int _roadOffset;
-  int _minRoadEdgeWidth;
-  int _minRoadWidth;
-  int _maxRoadWidth;
-  Vec _startPosition;
+  // Vec _startPosition;
+  List<Vec> _startingPositions = [];
 
-  Vec get startPosition => _startPosition;
-
-  bool _bendingLeft = false;
-  bool _bendingRight = false;
-  bool _widening = false;
-  bool _narrowing = false;
-
-  bool get roadBendingLeft => _bendingLeft;
-  bool get roadBendingRight => _bendingRight;
-  bool get roadWidening => _widening;
-  bool get roadNarrowing => _narrowing;
-
-  void toggleBendingLeft() {
-    _bendingLeft = !_bendingLeft;
-  }
-
-  void toggleBendingRight() {
-    _bendingRight = !_bendingRight;
-  }
-
-  void toggleWidening() {
-    _widening = !_widening;
-  }
-
-  void toggleNarrowing() {
-    _narrowing = !_narrowing;
-  }
+  // Vec get startPosition => _startPosition;
+  List<Vec> get startingPositions => _startingPositions;
 
   int get width => _width;
   int get height => _height;
-
-  int get leftEdge => _roadOffset;
-  int get rightEdge => leftEdge + _roadWidth - 1;
-  bool get roomLeft => leftEdge > _minRoadEdgeWidth;
-  bool get roomRight => rightEdge < width - _minRoadEdgeWidth - 1;
-  bool get roadAtMinWidth => _roadWidth <= _minRoadWidth;
-  bool get roadAtMaxWidth => _roadWidth >= _maxRoadWidth;
-  bool get roadWide => _roadWidth > _minRoadWidth;
-  int get roomOnLeft => leftEdge - _minRoadEdgeWidth;
-  int get roomOnRight => width - rightEdge - _minRoadEdgeWidth - 1;
 
   final math.Random _random = math.Random();
 
@@ -66,17 +27,21 @@ class Track {
 
   double get randomDouble => _random.nextDouble();
 
-  Track(int foo, int bar) {
+  Track() {
     initialize();
   }
 
   void initialize() {
-    _minRoadEdgeWidth = 2;
-    _minRoadWidth = 14;
-    _maxRoadWidth = width - _minRoadEdgeWidth * 2;
-    _roadWidth = _maxRoadWidth;
-    _roadOffset = _minRoadEdgeWidth;
-    _startPosition = Vec(rightEdge - 2, height - 5);
+    // _startPosition = Vec(23, 28);
+
+    _startingPositions = [
+      Vec(14, 20),
+      Vec(16, 20),
+      Vec(12, 22),
+      Vec(18, 22),
+      Vec(10, 24),
+      Vec(20, 24),
+    ];
 
     var roadSection = Road.wideStraight.first;
 
@@ -93,138 +58,11 @@ class Track {
     }
   }
 
-  void narrowRoad() {
-    if (_roadWidth > _minRoadWidth) {
-      _roadWidth -= 2;
-      if (roomRight && randomBool) {
-        rollRoadRight();
-      }
-    }
-  }
-
-  void onNarrowing() {
-    narrowRoad();
-  }
-
-  void widenRoad() {
-    if (_roadWidth < _maxRoadWidth) {
-      _roadWidth += 2;
-      if (roomLeft && randomBool) {
-        rollRoadLeft();
-      }
-    }
-  }
-
-  void onWidening() {
-    widenRoad();
-    if (!roomRight) {
-      rollRoadLeft();
-    }
-  }
-
-  void rollRoadLeft() {
-    if (roomLeft) {
-      _roadOffset -= 1;
-    }
-  }
-
-  void onBendingLeft() {
-    rollRoadLeft();
-  }
-
-  void rollRoadRight() {
-    if (roomRight) {
-      _roadOffset += 1;
-    }
-  }
-
-  void onBendingRight() {
-    rollRoadRight();
-  }
-
-  int _counter = 0;
-
-  bool get countdownDone => _counter <= 0;
-
-  void setCountdown([int nrOfTurns]) {
-    _counter = nrOfTurns ?? randomInt(4, 12);
-  }
-
-  void countdown() {
-    if (_counter > 0) {
-      _counter -= 1;
-    }
-  }
-
-  void resetRoadChanges() {
-    _bendingLeft = false;
-    _bendingRight = false;
-    _narrowing = false;
-    _widening = false;
-  }
-
-  void update([int numberOfRolldowns = 1]) {
-    for (var i = 0; i < numberOfRolldowns; i++) {
-      countdown();
-    }
-    if (countdownDone) {
-      resetRoadChanges();
-
-      if (randomBool) {
-        setCountdown();
-
-        if (roadWide && randomBool) {
-          toggleNarrowing();
-        } else if (roomOnLeft > roomOnRight || (randomBool && roomLeft)) {
-          toggleBendingLeft();
-        } else if (roomOnRight > roomOnLeft || (randomBool && roomRight)) {
-          toggleBendingRight();
-        } else if (randomBool && !roadAtMinWidth) {
-          toggleNarrowing();
-        }
-      }
-    }
-
-    if (roadWidening) {
-      onWidening();
-      if (countdownDone) {
-        toggleWidening();
-      }
-    } else if (roadNarrowing) {
-      onNarrowing();
-      if (countdownDone) {
-        toggleNarrowing();
-      }
-    } else if (roadBendingLeft) {
-      onBendingLeft();
-      if (countdownDone) {
-        toggleBendingLeft();
-      }
-    } else if (roadBendingRight) {
-      onBendingRight();
-      if (countdownDone) {
-        toggleBendingRight();
-      }
-    }
-  }
+  void update([int numberOfRolldowns = 1]) {}
 
   String generateRoadSection() {
     rollRoad();
     return _currentSection.last;
-
-    // var roadSection = '';
-    // var char;
-    // for (var x = 0; x < width; x++) {
-    //   if (x < leftEdge || x > rightEdge) {
-    //     char = '#';
-    //   } else {
-    //     char = '.';
-    //   }
-
-    //   roadSection += char;
-    // }
-
-    // return roadSection;
   }
 
   bool outOfBounds(Vec p) {
@@ -262,30 +100,64 @@ class Track {
     }
   }
 
-  void renderDebugInfo(Terminal terminal) {
-    terminal.writeAt(2, 2, '_bendingLeft: $_bendingLeft');
-    terminal.writeAt(2, 3, '_bendingRight: $_bendingRight');
-    terminal.writeAt(2, 4, '_widening: $_widening');
-    terminal.writeAt(2, 5, '_narrowing: $_narrowing');
-    terminal.writeAt(2, 6, '_counter: $_counter');
-    // terminal.writeAt(2, 7, 'currentWidth: $currentWidth');
-    terminal.writeAt(2, 8, '_roadOffset: $_roadOffset');
-    terminal.writeAt(2, 9, '_roadWidth: $_roadWidth');
-    terminal.writeAt(2, 10, '_minRoadWidth: $_minRoadWidth');
-    terminal.writeAt(2, 11, '_maxRoadWidth: $_maxRoadWidth');
-    // terminal.writeAt(2, 12, 'atMaxWidth: $atMaxWidth');
-    terminal.writeAt(2, 13, 'leftEdge: $leftEdge');
-    terminal.writeAt(2, 14, 'rightEdge: $rightEdge');
-    terminal.writeAt(2, 15, 'roomOnLeft: $roomOnLeft');
-    terminal.writeAt(2, 16, 'roomOnRight: $roomOnRight');
-  }
+  void renderDebugInfo(Terminal terminal) {}
 
   final List<String> _currentSection = List.empty(growable: true);
   final List<String> _nextSection = List.empty(growable: true);
 
+  RoadStyle _roadStyle = RoadStyle.wide;
+
+  bool get roadIsWide => _roadStyle == RoadStyle.wide;
+  bool get roadIsNarrowCenter => _roadStyle == RoadStyle.narrowCenter;
+  bool get roadIsNarrowRight => _roadStyle == RoadStyle.narrowRight;
+
+  void setRoadNarrowCenter() {
+    _roadStyle = RoadStyle.narrowCenter;
+  }
+
+  void setRoadNarrowRight() {
+    _roadStyle = RoadStyle.narrowRight;
+  }
+
+  void setRoadWide() {
+    _roadStyle = RoadStyle.wide;
+  }
+
   List<String> generateRoad() {
     var roadSections = [];
-    roadSections.addAll(Road.wideSections);
+
+    if (roadIsWide && randomBool) {
+      setRoadNarrowCenter();
+      return Road.wideToNarrowCenter;
+    } else if (roadIsNarrowCenter && randomBool) {
+      if (randomBool) {
+        setRoadNarrowRight();
+        return Road.narrowCenterToNarrowRight;
+      } else {
+        setRoadWide();
+        return Road.narrowCenterToWide;
+      }
+    } else if (roadIsNarrowRight && randomBool) {
+      if (randomBool) {
+        setRoadNarrowCenter();
+        return Road.narrowRightToNarrowCenter;
+      }
+    }
+
+    switch (_roadStyle) {
+      case RoadStyle.wide:
+        roadSections.addAll(Road.wideSections);
+        break;
+      case RoadStyle.narrowCenter:
+        roadSections.addAll(Road.narrowCenterSections);
+        break;
+      case RoadStyle.narrowRight:
+        roadSections.addAll(Road.narrowRightSections);
+        break;
+      default:
+        roadSections.addAll(Road.wideSections);
+    }
+
     var randomNumber = randomInt(0, roadSections.length);
     return roadSections[randomNumber];
   }
@@ -293,6 +165,7 @@ class Track {
   void initializeNewRoad() {
     _currentSection.clear();
     _nextSection.clear();
+    _roadStyle = RoadStyle.wide;
     _currentSection.addAll(Road.wideStraight);
     _nextSection.addAll(generateRoad());
   }